now using modelreader instead of modelcoordinator
